src/perllib/utils.pl: more get_line_field perf enhancements
Turns out pointer arithmetic + substr is faster than the native split.
The argument here is that split is written in c and therefore much
faster than anything that perl code could do. That doesn't seem to be
the case though. Building the array natively is more expensive than
the perl code to do pointer manip.


git-svn-id: http://crush-tools.googlecode.com/svn/trunk@388 9f305d85-0c50-0410-b3e0-65f4174694de